requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
python-dotenv
google-generativeai